feat: rework formatters
